+++ CHANGELOG.md
@@ -14,6 +14,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### Fixes
 
+- [#2348](https://gitlab.com/meltano/meltano/-/issues/2348) Invalidate `meltano select` catalog discovery cache when extractor configuration is changed
+
 ### Breaks